Rename getDataChartPie to getDefaultDate because fetch is now on App
