+++ src/App.js
@@ -5,7 +5,7 @@ import NewCall from "./components/NewCall";
 import CallHistory from "./components/CallHistory";
 import { getData } from "./services/getData";
 import { getList } from "./services/getList";
-import { currentTime } from "./services/getDataChartPie";
+import { currentTime } from "./services/getDefaultDate";
 import "./styles/App.scss";
 import { Route, Switch } from "react-router-dom";
 import Modal from "./components/Modal";
